chore: move commands to cli folder
Fixes #53
