+++ src/autotarcompress/cli/commands/command.py
