+++ src/autotarcompress/cli/commands/info.py
@@ -6,7 +6,7 @@ the last backup operation.
 
 import logging
 
-from autotarcompress.commands.command import Command
+from autotarcompress.cli.commands.command import Command
 from autotarcompress.config import BackupConfig
 from autotarcompress.info_manager import InfoManager
